Cache line number so importing settins doesn't take ~500ms
